pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "confluence_webapp"
version = "1.0.0"
description = "Offers a web interface for cell confluence measurements with the model Cellpose, Detectron2, U-Net, and Segment Anything."
authors = [
    { name = "Max Joas", email = "joas@informatik.uni-leipzig.de" }
]
dependencies = [
    "requests>=2.28.0",
    "pandas==2.2.2",
    "streamlit==1.26.0",
    "torchvision==0.14.0",
    "Pillow==9.4.0",
    "scipy==1.10.1",
    "opencv-python==4.7.0.68",
    "opencv-python-headless==4.7.0.72",
    "matplotlib==3.6.2",
    "munch==4.0.0",
    "numpy==1.24.2",
    "seaborn==0.12.2",
    "scikit-image==0.19.3",
    "scikit-learn==1.2.2",
    "scikit-multilearn==0.2.0",
    "toml==0.10.2",
    "torch==1.13.0",
    "pip==22.0.4",
    "pycocotools==2.0.6",
    "cellpose @ git+https://github.com/MaxJoas/cellpose_fork@main",
    "detectron2 @ git+https://github.com/facebookresearch/detectron2.git@650c60ed365db1b04765ffd121574c0765a5c24c",
    "segment_anything==1.0",
]

requires-python = ">=3.8"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=3.0",
]

[tool.pytest.ini_options]
addopts = "--cov=src"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src"]  # Include the src directory

[tool.hatch.metadata]
allow-direct-references = true